add nuvla.api, add notifs on app published, refactor 'matching' package, update models and notifications, add tests
