+++ code/requirements.txt
@@ -1,3 +1,4 @@
 kafka-python==2.0.2
 elasticsearch==7.11.0
 python-dateutil
+nuvla-api==3.0.9
